@@ -1,6 +1,7 @@
-use crate::clash::{headers, MinreqResult};
-
-use super::{ClashUtil, TIMEOUT};
+use super::{
+    super::{MinreqResult, headers},
+    ClashUtil,
+};
 
 impl ClashUtil {
     pub fn mock_clash_core<U: Into<minreq::URL>>(&self, url: U, with_proxy: bool) -> MinreqResult {
@@ -8,6 +9,7 @@ impl ClashUtil {
             url,
             with_proxy.then_some(self.proxy_addr.as_str()),
             Some(self.ua.as_deref().unwrap_or("clash.meta")),
+            self.timeout,
         )
     }
     pub fn dl_github<U: Into<minreq::URL>>(
@@ -21,7 +23,7 @@ impl ClashUtil {
             req = req.with_proxy(minreq::Proxy::new(&self.proxy_addr)?);
         }
         req = req.with_header(headers::AUTHORIZATION, format!("Bearer {token}"));
-        req.with_timeout(*TIMEOUT).send_lazy()
+        req.with_timeout(self.timeout).send_lazy()
     }
     pub fn dl_gitlab<U: Into<minreq::URL>>(
         &self,
@@ -34,7 +36,7 @@ impl ClashUtil {
             req = req.with_proxy(minreq::Proxy::new(&self.proxy_addr)?);
         }
         req = req.with_header("PRIVATE-TOKEN", token);
-        req.with_timeout(*TIMEOUT).send_lazy()
+        req.with_timeout(self.timeout).send_lazy()
     }
 }
 
