@@ -1,12 +1,8 @@
-use super::{ipc, BackEnd};
+use super::{BackEnd, Mode, State, clash::webapi::ClashConfig, ipc};
 
 #[allow(unused_imports)] // currently, only [`SwitchMode`] is impl on macOS
 use ipc::exec;
 use std::io::Error;
-
-mod state;
-
-pub(super) use state::State;
 
 #[cfg_attr(
     feature = "tui",
@@ -32,19 +28,64 @@ pub enum ServiceOp {
 }
 
 impl BackEnd {
-    pub fn update_mode(&self, mode: crate::clash::webapi::Mode) -> anyhow::Result<()> {
+    pub fn restart_clash(&self) -> std::io::Result<String> {
+        self.api.restart(None).map_err(Error::other)
+    }
+    pub fn update_state(
+        &self,
+        new_pf: Option<String>,
+        new_mode: Option<Mode>,
+    ) -> anyhow::Result<State> {
+        if let Some(mode) = new_mode {
+            self.update_mode(mode)?;
+        }
+        if let Some(pf) = new_pf.as_ref() {
+            if let Some(pf) = self.as_profile().get_profile(pf) {
+                self.as_profile().select_profile(pf)?;
+            } else {
+                anyhow::bail!("Not a recorded profile");
+            };
+        }
+        #[cfg(target_os = "windows")]
+        let sysproxy = self.is_system_proxy_enabled().map_or_else(
+            |v| {
+                log::error!("{}", v);
+                None
+            },
+            Some,
+        );
+        let ClashConfig { mode, tun, .. } = self.api.config_get()?;
+        Ok(State {
+            profile: new_pf.unwrap_or(self.as_profile().get_current_profile().name),
+            mode: Some(mode),
+            tun: if tun.enable { Some(tun.stack) } else { None },
+            #[cfg(target_os = "windows")]
+            sysproxy,
+        })
+    }
+    pub fn get_clash_version(&self) -> anyhow::Result<String> {
+        let v = self.api.version()?;
+        let map: serde_json::Value = serde_json::from_str(&v).unwrap();
+        Ok(map
+            .get("version")
+            .and_then(|v| v.as_str())
+            .unwrap_or("v0.0.0")
+            .to_owned())
+    }
+    pub fn update_mode(&self, mode: Mode) -> anyhow::Result<()> {
         let load = format!(r#"{{"mode": "{mode}"}}"#);
         self.api.config_patch(load)?;
         Ok(())
     }
+
     #[cfg(target_os = "linux")]
     pub fn clash_srv_ctl(&self, op: ServiceOp) -> Result<String, Error> {
         match op {
             ServiceOp::RestartClashService => {
                 let arg = if self.cfg.service.is_user {
-                    vec!["--user", self.cfg.service.clash_srv_nam.as_str()]
+                    vec!["--user", self.cfg.service.clash_service_name.as_str()]
                 } else {
-                    vec![self.cfg.service.clash_srv_nam.as_str()]
+                    vec![self.cfg.service.clash_service_name.as_str()]
                 };
                 {
                     let mut args = vec!["restart"];
@@ -60,9 +101,9 @@ impl BackEnd {
             ServiceOp::RestartClashCore => self.restart_clash(),
             ServiceOp::StopClashService => {
                 let arg = if self.cfg.service.is_user {
-                    vec!["--user", self.cfg.service.clash_srv_nam.as_str()]
+                    vec!["--user", self.cfg.service.clash_service_name.as_str()]
                 } else {
-                    vec![self.cfg.service.clash_srv_nam.as_str()]
+                    vec![self.cfg.service.clash_service_name.as_str()]
                 };
                 {
                     let mut args = vec!["stop"];
@@ -76,12 +117,12 @@ impl BackEnd {
                 }
             }
             ServiceOp::SetPermission => {
-                exec("chmod", vec!["+x", self.cfg.basic.clash_bin_pth.as_str()])?;
+                exec("chmod", vec!["+x", self.cfg.basic.clash_bin_path.as_str()])?;
                 ipc::exec_with_sbin(
                     "setcap",
                     vec![
                         "'cap_net_admin,cap_net_bind_service=+ep'",
-                        self.cfg.basic.clash_bin_pth.as_str(),
+                        self.cfg.basic.clash_bin_path.as_str(),
                     ],
                 )
             }
@@ -113,12 +154,12 @@ impl BackEnd {
             ServiceOp::RestartClashService => {
                 start_process_as_admin(
                     NSSM_PGM,
-                    format!("restart {}", self.cfg.service.clash_srv_nam).as_str(),
+                    format!("restart {}", self.cfg.service.clash_service_name).as_str(),
                     true,
                 )?;
                 exec(
                     NSSM_PGM,
-                    vec!["status", self.cfg.service.clash_srv_nam.as_str()],
+                    vec!["status", self.cfg.service.clash_service_name.as_str()],
                 )
             }
 
@@ -127,12 +168,12 @@ impl BackEnd {
             ServiceOp::StopClashService => {
                 start_process_as_admin(
                     NSSM_PGM,
-                    &format!("stop {}", self.cfg.service.clash_srv_nam),
+                    &format!("stop {}", self.cfg.service.clash_service_name),
                     true,
                 )?;
                 exec(
                     NSSM_PGM,
-                    vec!["status", self.cfg.service.clash_srv_nam.as_str()],
+                    vec!["status", self.cfg.service.clash_service_name.as_str()],
                 )
             }
 
@@ -141,7 +182,7 @@ impl BackEnd {
                     NSSM_PGM,
                     &format!(
                         "install {} \"{}\" -d \"{}\" -f \"{}\"",
-                        self.cfg.service.clash_srv_nam,
+                        self.cfg.service.clash_service_name,
                         self.cfg.basic.clash_bin_pth,
                         self.cfg.basic.clash_cfg_dir,
                         self.cfg.basic.clash_cfg_pth
@@ -151,14 +192,14 @@ impl BackEnd {
 
                 exec(
                     NSSM_PGM,
-                    vec!["status", self.cfg.service.clash_srv_nam.as_str()],
+                    vec!["status", self.cfg.service.clash_service_name.as_str()],
                 )
             }
 
             ServiceOp::UnInstallSrv => ipc::execute_powershell_script_as_admin(
                 &format!(
                     "{0} stop {1}; {0} remove {1}",
-                    NSSM_PGM, self.cfg.service.clash_srv_nam
+                    NSSM_PGM, self.cfg.service.clash_service_name
                 ),
                 true,
             ),
@@ -197,53 +238,5 @@ impl BackEnd {
     #[cfg(windows)]
     pub fn is_system_proxy_enabled(&self) -> std::io::Result<bool> {
         ipc::is_system_proxy_enabled()
-    }
-}
-
-impl BackEnd {
-    pub fn restart_clash(&self) -> std::io::Result<String> {
-        self.api.restart(None).map_err(Error::other)
-    }
-    pub fn update_state(
-        &self,
-        new_pf: Option<String>,
-        new_mode: Option<crate::clash::webapi::Mode>,
-    ) -> anyhow::Result<State> {
-        use crate::clash::webapi::ClashConfig;
-        if let Some(mode) = new_mode {
-            self.update_mode(mode)?;
-        }
-        if let Some(pf) = new_pf.as_ref() {
-            if let Some(pf) = self.get_profile(pf) {
-                self.select_profile(pf)?;
-            } else {
-                anyhow::bail!("Not a recorded profile");
-            };
-        }
-        #[cfg(target_os = "windows")]
-        let sysp = self.is_system_proxy_enabled().map_or_else(
-            |v| {
-                log::error!("{}", v);
-                None
-            },
-            Some,
-        );
-        let ClashConfig { mode, tun, .. } = self.api.config_get()?;
-        Ok(State {
-            profile: new_pf.unwrap_or(self.get_current_profile().name),
-            mode: Some(mode),
-            tun: if tun.enable { Some(tun.stack) } else { None },
-            #[cfg(target_os = "windows")]
-            sysproxy: sysp,
-        })
-    }
-    pub fn get_clash_version(&self) -> Result<String, String> {
-        let v = self.api.version().map_err(|e| e.to_string())?;
-        let map: serde_json::Value = serde_json::from_str(&v).unwrap();
-        Ok(map
-            .get("version")
-            .and_then(|v| v.as_str())
-            .unwrap_or("v0.0.0")
-            .to_owned())
     }
 }
