@@ -1,44 +1,42 @@
 #![warn(clippy::all)]
 #![deny(unsafe_code)]
-#![warn(keyword_idents_2024)]
+
 mod backend;
-mod clash;
-mod commands;
+mod cli;
 #[cfg(feature = "tui")]
 mod tui;
 mod utils;
 
 use backend::BackEnd;
-use utils::{consts, BuildConfig};
+use utils::{BuildConfig, consts};
 
 /// The clashtui config dir
+///
+/// Access relative path via [`crate::consts`]
 struct DataDir;
 
 fn main() {
     if is_root::is_root() {
         println!("{}", consts::ROOT_WARNING)
     }
-    let Ok((infos, verbose_level)) = commands::parse_args() else {
+    let Ok((args, verbose_level)) = cli::parse_args() else {
         return;
     };
     let backend = match BuildConfig::load_config() {
         Ok(v) => BackEnd::build(v),
-        Err(e) => reinit_config_dir(e),
+        Err(e) => interactive_reinit_config_dir(e),
     };
-    utils::setup_logging(verbose_level);
+    utils::logging::setup(verbose_level);
     // handle commands
-    if let Err(e) = match infos {
-        Some(command) => commands::handle_cli(command, backend),
+    if let Err(e) = match args {
+        Some(command) => cli::handle_cli(command, backend),
         #[cfg(not(feature = "tui"))]
         None => {
             eprintln!("use `--help/-h` for help");
             Ok(())
         }
         #[cfg(feature = "tui")]
-        None => {
-            println!("Entering TUI...");
-            start_tui(backend)
-        }
+        None => start_tui(backend),
     } {
         eprintln!("clashtui encounter some error: {e}");
         log::error!("Err: {e}");
@@ -57,12 +55,15 @@ async fn start_tui(backend: BackEnd) -> anyhow::Result<()> {
     setup::setup()?;
     // and recovery from it when panic
     setup::set_panic_hook();
+
     let frontend = tui::FrontEnd::new();
-    let (backend_tx, app_rx) = tokio::sync::mpsc::channel(5);
-    let (app_tx, backend_rx) = tokio::sync::mpsc::channel(5);
+    let (backend_tx, frontend_rx) = tokio::sync::mpsc::channel(5);
+    let (frontend_tx, backend_rx) = tokio::sync::mpsc::channel(5);
+
     let backend = tokio::spawn(backend.run(backend_tx, backend_rx));
-    let frontend = tokio::spawn(frontend.run(app_tx, app_rx));
+    let frontend = tokio::spawn(frontend.run(frontend_tx, frontend_rx));
     let (frontend, backend) = tokio::try_join!(frontend, backend)?;
+
     setup::restore()?;
     // clear the result, save profiles to disk
     frontend?;
@@ -77,9 +78,9 @@ async fn start_tui(backend: BackEnd) -> anyhow::Result<()> {
 ///
 /// # Panics
 /// if unable to write/read from stdio
-fn reinit_config_dir(err: impl std::fmt::Display) -> ! {
+fn interactive_reinit_config_dir(err: impl std::fmt::Display) -> ! {
     // we don't really do so, as it can be dangerous
-    if commands::Confirm::default()
+    if cli::Confirm::default()
         .append_prompt(format!("failed to load config: {err}"))
         .append_prompt("if you are upgrading from old version, please run `clashtui migrate`")
         .append_prompt("program will try to init default config")
