@@ -1,6 +1,6 @@
 pub const TAB_TITLE_SERVICE: &str = "ServiceCTL";
 pub const TAB_TITLE_PROFILE: &str = "ProfileCTL";
-#[cfg(feature = "connection-tab")]
+#[cfg(feature = "connections")]
 pub const TAB_TITLE_CONNECTION: &str = "ConnectionCTL";
 #[cfg(feature = "template")]
 pub const TAB_TITLE_TEMPLATE: &str = "TemplateCTL";
